adding colour circle to user
