+++ app/client/src/components/RequireLayout.jsx
@@ -5,8 +5,7 @@ export default function RequireLayout() {
 
     const { auth } = useAuth();
     const location = useLocation();
-
-    // console.log(auth);
+    console.log(auth);
 
     return (
         auth ?
