+++ app/client/src/context/AuthProvider.jsx
@@ -5,8 +5,6 @@ const AuthContext = createContext({});
 
 export const AuthProvider = ({ children }) => { // provides the authentication to the different components
     const [auth, setAuth] = useLocalStorage('user', null);
-    console.log(auth);
-    console.log(setAuth);
     const [players, setPlayers] = useState();
     const [player, setPlayer] = useState({});
